EW-287 - Split keycloak initialization and configuration (#3570)
Splits initialization and configuration of ErWIn IDM (Keycloak) for the different environments (local, cluster - dev, prod).
