+++ apps/server/src/modules/management/management.module.ts
@@ -4,7 +4,7 @@ import { ConfigModule } from '@nestjs/config';
 import { ConsoleWriterService } from '@shared/infra/console';
 import { DatabaseManagementModule, DatabaseManagementService } from '@shared/infra/database';
 import { FileSystemModule } from '@shared/infra/file-system';
-import { KeycloakModule } from '@shared/infra/identity-management/keycloak/keycloak.module';
+import { KeycloakControllerModule } from '@shared/infra/identity-management/keycloak/controller/keycloak.controller.module';
 import serverConfig from '@src/server.config';
 import { DatabaseManagementConsole } from './console/database-management.console';
 import { DatabaseManagementController } from './controller/database-management.controller';
@@ -22,7 +22,7 @@ const baseImports = [
 ];
 
 const imports = (Configuration.get('FEATURE_IDENTITY_MANAGEMENT_ENABLED') as boolean)
-	? [...baseImports, KeycloakModule]
+	? [...baseImports, KeycloakControllerModule]
 	: baseImports;
 
 const providers = [
