+++ apps/server/src/shared/infra/identity-management/identity-management.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { IdentityManagementService } from './identity-management.service';
 import { KeycloakModule } from './keycloak/keycloak.module';
-import { KeycloakIdentityManagementService } from './keycloak/keycloak-identity-management.service';
+import { KeycloakIdentityManagementService } from './keycloak/service/keycloak-identity-management.service';
 
 @Module({
 	imports: [KeycloakModule],
