+++ apps/server/src/shared/infra/identity-management/keycloak/keycloak.module.ts
@@ -1,19 +1,18 @@
 import { Module } from '@nestjs/common';
-import { ConsoleWriterModule } from '@shared/infra/console';
 import { LoggerModule } from '@src/core/logger';
 import { SystemRepo } from '@shared/repo';
 import KeycloakAdminClient from '@keycloak/keycloak-admin-client';
-import { KeycloakConsole } from './console/keycloak-management.console';
-import { KeycloakAdministrationService } from './keycloak-administration.service';
 import { KeycloakSettings } from './interface/keycloak-settings.interface';
 import { KeycloakManagementUc } from './uc/Keycloak-management.uc';
-import { KeycloakManagementController } from './controller/keycloak-management.controller';
 import { KeycloakManagementInputFiles } from './interface';
 import KeycloakConfiguration from './keycloak-config';
+import { KeycloakAdministrationService } from './service/keycloak-administration.service';
+import { KeycloakConfigurationService } from './service/keycloak-configuration.service';
+import { KeycloakSeedService } from './service/keycloak-seed.service';
 
 @Module({
-	imports: [ConsoleWriterModule, LoggerModule],
-	controllers: [KeycloakManagementController],
+	imports: [LoggerModule],
+	controllers: [],
 	providers: [
 		SystemRepo,
 		KeycloakAdminClient,
@@ -27,8 +26,9 @@ import KeycloakConfiguration from './keycloak-config';
 			useValue: KeycloakConfiguration.keycloakInputFiles,
 		},
 		KeycloakManagementUc,
-		KeycloakConsole,
+		KeycloakConfigurationService,
+		KeycloakSeedService,
 	],
-	exports: [KeycloakAdministrationService, KeycloakConsole],
+	exports: [KeycloakAdministrationService, KeycloakManagementUc],
 })
 export class KeycloakModule {}
