+++ apps/server/src/shared/infra/identity-management/keycloak/service/keycloak-identity-management.service.spec.ts
@@ -1,8 +1,8 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import KeycloakAdminClient from '@keycloak/keycloak-admin-client';
 import { KeycloakIdentityManagementService } from './keycloak-identity-management.service';
-import { IdentityManagementService } from '../identity-management.service';
-import { KeycloakSettings } from './interface/keycloak-settings.interface';
+import { IdentityManagementService } from '../../identity-management.service';
+import { KeycloakSettings } from '../interface/keycloak-settings.interface';
 import { KeycloakAdministrationService } from './keycloak-administration.service';
 
 describe('KeycloakIdentityManagement', () => {
