+++ apps/server/src/shared/infra/identity-management/keycloak/service/keycloak-identity-management.service.ts
@@ -3,7 +3,7 @@ import { IAccount, IAccountUpdate } from '@shared/domain';
 /* eslint-disable no-nested-ternary */
 import { Injectable } from '@nestjs/common';
 import UserRepresentation from '@keycloak/keycloak-admin-client/lib/defs/userRepresentation';
-import { IdentityManagementService } from '../identity-management.service';
+import { IdentityManagementService } from '../../identity-management.service';
 import { KeycloakAdministrationService } from './keycloak-administration.service';
 
 @Injectable()
