+++ package.json
@@ -29,7 +29,8 @@
 		"setup": "npm run setup:db",
 		"setup:db": "npm run nest:start:console -- database seed",
 		"setup:idm:seed": "npm run nest:start:console -- idm seed",
-		"setup:idm:configure": "npm run nest:start:console -- idm configure",
+		"setup:idm:configure": "npm run nest:start:console -- idm configure --from-json true",
+		"setup:idm": "npm run setup:idm:seed && npm run setup:idm:configure",
 		"build": "npm run nest:build",
 		"copy-files": "copyfiles src/**/*.yaml build",
 		"inspect": "npm run build && node --inspect=5959 build/src/",
